Finalisation crud catégorie
-Les users normaux peuvent voir les catégories, et le détails d'une catégorie
