+++ .idea/workspace.xml
@@ -20,13 +20,18 @@
   </component>
   <component name="ChangeListManager">
     <list default="true" id="9e8fa817-45b1-4092-bf2e-242d5d3a7099" name="Default Changelist" comment="">
+      <change afterPath="$PROJECT_DIR$/templates/admin_user/new.html.twig" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/src/Controller/Admin/AdminUserController.php" beforeDir="false" afterPath="$PROJECT_DIR$/src/Controller/Admin/AdminUserController.php" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/src/Controller/Admin/CategoryController.php" beforeDir="false" afterPath="$PROJECT_DIR$/src/Controller/Admin/CategoryController.php" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/Controller/RegistrationController.php" beforeDir="false" afterPath="$PROJECT_DIR$/src/Controller/RegistrationController.php" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/Controller/HomeController.php" beforeDir="false" afterPath="$PROJECT_DIR$/src/Controller/HomeController.php" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/Controller/RegistrationController.php" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/templates/admin_user/edit.html.twig" beforeDir="false" afterPath="$PROJECT_DIR$/templates/admin_user/edit.html.twig" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/templates/admin_user/index.html.twig" beforeDir="false" afterPath="$PROJECT_DIR$/templates/admin_user/index.html.twig" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/templates/advert/index.html.twig" beforeDir="false" afterPath="$PROJECT_DIR$/templates/advert/index.html.twig" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/templates/category/index.html.twig" beforeDir="false" afterPath="$PROJECT_DIR$/templates/category/index.html.twig" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/templates/category/show.html.twig" beforeDir="false" afterPath="$PROJECT_DIR$/templates/category/show.html.twig" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/templates/registration/register.html.twig" beforeDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -161,13 +166,17 @@
   <component name="PropertiesComponent">
     <property name="RunOnceActivity.OpenProjectViewOnStart" value="true" />
     <property name="WebServerToolWindowFactoryState" value="false" />
-    <property name="last_opened_file_path" value="$PROJECT_DIR$" />
+    <property name="last_opened_file_path" value="$PROJECT_DIR$/templates/admin_user" />
     <property name="nodejs_package_manager_path" value="npm" />
     <property name="settings.editor.selected.configurable" value="Errors" />
     <property name="vue.rearranger.settings.migration" value="true" />
   </component>
   <component name="RecentsManager">
+    <key name="CopyFile.RECENT_KEYS">
+      <recent name="C:\wamp64\www\lebonangle\templates\admin_user" />
+    </key>
     <key name="MoveFile.RECENT_KEYS">
+      <recent name="C:\wamp64\www\lebonangle\templates\admin_user" />
       <recent name="C:\wamp64\www\lebonangle\src\Controller\Admin" />
       <recent name="C:\wamp64\www\lebonangle\src\Controller" />
     </key>
@@ -183,7 +192,7 @@
       <workItem from="1605978192686" duration="1913000" />
       <workItem from="1606124501760" duration="5361000" />
       <workItem from="1606203988231" duration="24959000" />
-      <workItem from="1606289966494" duration="10727000" />
+      <workItem from="1606289966494" duration="15188000" />
     </task>
     <servers />
   </component>
@@ -207,10 +216,10 @@
       <screen x="0" y="0" width="1920" height="1040" />
     </state>
     <state x="122" y="0" key="#com.intellij.refactoring.rename.AutomaticRenamingDialog/1920.-213.2560.1400/0.0.1920.1040@0.0.1920.1040" timestamp="1606127552686" />
-    <state x="690" y="274" key="#com.jetbrains.php.actions.copyPaste.importReferences.PhpImportReferencesDialog" timestamp="1606298712015">
+    <state x="690" y="274" key="#com.jetbrains.php.actions.copyPaste.importReferences.PhpImportReferencesDialog" timestamp="1606311068003">
       <screen x="0" y="0" width="1920" height="1040" />
     </state>
-    <state x="690" y="274" key="#com.jetbrains.php.actions.copyPaste.importReferences.PhpImportReferencesDialog/1920.-213.2560.1400/0.0.1920.1040@0.0.1920.1040" timestamp="1606298712015" />
+    <state x="690" y="274" key="#com.jetbrains.php.actions.copyPaste.importReferences.PhpImportReferencesDialog/1920.-213.2560.1400/0.0.1920.1040@0.0.1920.1040" timestamp="1606311068003" />
     <state x="740" y="274" key="FileChooserDialogImpl" timestamp="1606212950337">
       <screen x="0" y="0" width="1920" height="1040" />
     </state>
