+++ templates/admin_user/_form.html.twig
@@ -1,0 +1,5 @@
+{{ form_start(form) }}
+        {{ form_widget(form) }}
+        <button class="btn btn-success">{{ button_label|default('Valider') }}</button>
+        <a class="btn btn-secondary" href="{{ path('admin_user_index') }}">Retour</a>
+{{ form_end(form) }}
