+++ templates/admin_user/edit.html.twig
@@ -9,11 +9,7 @@
 
     <hr>
 
-    {{ form_start(form) }}
-        {{ form_widget(form) }}
-        <button class="btn btn-success">{{ button_label|default('Valider') }}</button>
-        <a class="btn btn-secondary" href="{{ path('admin_user_index') }}">Retour</a>
-    {{ form_end(form) }}
+    {{ include('admin_user/_form.html.twig') }}
 
 
 {% endblock %}
