+++ templates/admin_user/index.html.twig
@@ -49,5 +49,5 @@
         </tbody>
     </table>
 
-    <a class="btn btn-primary" href="{{ path('app_register') }}">Ajouter un Administrateur</a>
+    <a class="btn btn-primary" href="{{ path('admin_user_new') }}">Ajouter un Administrateur</a>
 {% endblock %}
