+++ templates/advert/index.html.twig
@@ -48,6 +48,10 @@
                 <td>{{ advert.content }}</td>
                 <td style="display: flex; justify-content: space-evenly;">
                     <a class="btn btn-primary" href="#"><i class="fas fa-eye"></i></a>
+                    {% if is_granted('ROLE_ADMIN') %}
+                        <a class="btn btn-primary" href="#"><i class="fas fa-edit"></i></a>
+                        <a class="btn btn-danger" href="#"><i class="fas fa-trash-alt"></i></a>
+                    {% endif %}
                 </td>
             </tr>
         {% else %}
