+++ templates/category/index.html.twig
@@ -22,12 +22,15 @@
                 <h5 class="card-title">{{category.name}}</h5>
                 <h6 class="card-subtitle mb-2 text-muted">#{{category.id}}</h6>
                 <div style="display: flex;">
-                    <a href="{{ path('category_edit', {'id': category.id}) }}" class="btn btn-primary"><i class="fas fa-edit"></i></a>
-                    <form method="post" action="{{ path('category_delete', {'id': category.id}) }}" onsubmit="return confirm('Êtes-vous certain de vouloir supprimer cet catégorie ?');">
-                        <input type="hidden" name="_method" value="DELETE">
-                        <input type="hidden" name="_token" value="{{ csrf_token('delete' ~ category.id) }}">
-                        <button class="btn btn-danger" style="margin-left: 5px;"><i class="fas fa-trash-alt"></i></button>
-                    </form>
+                    <a href="{{ path('category_show', {'id': category.id}) }}" class="btn btn-primary"><i class="fas fa-eye"></i></a>
+                    {% if is_granted('ROLE_ADMIN') %}
+                        <a href="{{ path('category_edit', {'id': category.id}) }}" class="btn btn-primary" style="margin-left: 5px;"><i class="fas fa-edit"></i></a>
+                        <form method="post" action="{{ path('category_delete', {'id': category.id}) }}" onsubmit="return confirm('Êtes-vous certain de vouloir supprimer cet catégorie ?');">
+                            <input type="hidden" name="_method" value="DELETE">
+                            <input type="hidden" name="_token" value="{{ csrf_token('delete' ~ category.id) }}">
+                            <button class="btn btn-danger" style="margin-left: 5px;"><i class="fas fa-trash-alt"></i></button>
+                        </form>
+                    {% endif %}
                 </div>
 
             </div>
